Move PickupDate, PickupDateSeries and Feedback models to new pickups app
