+++ foodsaving/pickups/tests/test_notifications.py
@@ -6,8 +6,9 @@ from django.utils import timezone
 from rest_framework.test import APITestCase
 
 from foodsaving.groups.factories import GroupFactory
-from foodsaving.stores.factories import StoreFactory, PickupDateFactory
+from foodsaving.stores.factories import StoreFactory
 from foodsaving.users.factories import UserFactory
+from pickups.factories import PickupDateFactory
 
 
 def parse(body):
