+++ foodsaving/pickups/tests/test_serializer.py
@@ -1,0 +1,1 @@
+# TODO: Write tests...
