+++ foodsaving/stores/admin.py
@@ -1,18 +1,8 @@
 from django.contrib import admin
 
-from foodsaving.stores.models import PickupDateSeries, PickupDate, Store
+from foodsaving.stores.models import Store
 
 
 @admin.register(Store)
 class StoreAdmin(admin.ModelAdmin):
-    pass
-
-
-@admin.register(PickupDate)
-class PickupDateAdmin(admin.ModelAdmin):
-    pass
-
-
-@admin.register(PickupDateSeries)
-class PickupDateSeriesAdmin(admin.ModelAdmin):
     pass
