+++ foodsaving/stores/tests/test_stores_api.py
@@ -1,13 +1,16 @@
 from copy import deepcopy
-from dateutil.parser import parse
-from dateutil.relativedelta import relativedelta
-from django.utils import timezone
+
 from rest_framework import status
 from rest_framework.test import APITestCase
+from dateutil.relativedelta import relativedelta
+from django.utils import timezone
+from dateutil.parser import parse
+
 from foodsaving.groups.factories import GroupFactory
-from foodsaving.stores.factories import StoreFactory, PickupDateSeriesFactory, PickupDateFactory
+from foodsaving.stores.factories import StoreFactory
 from foodsaving.users.factories import UserFactory
 from foodsaving.utils.tests.fake import faker
+from pickups.factories import PickupDateSeriesFactory, PickupDateFactory
 
 
 class TestStoresAPI(APITestCase):
@@ -151,6 +154,7 @@ class TestStoresAPI(APITestCase):
         self.assertEqual(len(response.data), 0)
 
 
+# TODO: Move the 'weeks_in_advance' field to PickupDateSeries; then also move this test class.
 class TestStoreChangesPickupDateSeriesAPI(APITestCase):
     @classmethod
     def setUpClass(cls):
