+++ foodsaving/userauth/tests/test_userauth_api.py
@@ -7,7 +7,8 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 
 from foodsaving.groups.factories import GroupFactory
-from foodsaving.stores.factories import StoreFactory, PickupDateFactory
+from foodsaving.stores.factories import StoreFactory
+from foodsaving.pickups.factories import PickupDateFactory
 from foodsaving.users.factories import UserFactory, VerifiedUserFactory
 from foodsaving.utils.tests.fake import faker
